moved all appendages to a single folder
