+++ uno/serial.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+picocom /dev/uno -b 115200 --echo
