+++ uno/src/uno.ino
@@ -1,0 +1,182 @@
+// Auto-generated by ArduinoGen
+
+// Includes
+
+#define STR1(x)  #x
+#define STR(x)  STR1(x)
+
+// Globals
+int ledState = HIGH;
+// Command parsing
+const int MAX_ARGS = 6;
+String args[MAX_ARGS];
+int numArgs = 0;
+
+// Pin definitions
+const char LED = 13;
+const char front_right_bumper_pin = 13;
+
+
+// Constructors
+const char front_right_bumper_index = 0;
+char switches[1] = {
+    front_right_bumper_pin
+};
+
+
+void setup() {
+    // Init LED pin
+    pinMode(LED, OUTPUT);
+
+    pinMode(front_right_bumper_pin, INPUT);
+
+    //Init Serial
+    Serial.begin(115200);
+}
+
+/* The loop is set up in two parts. First the Arduino does the work it needs to
+ * do for every loop, next is runs the checkInput() routine to check and act on
+ * any input from the serial connection.
+ */
+void loop() {
+    int inbyte;
+
+    // Accept and parse serial input
+    checkInput();
+}
+
+void parse_args(String command) {
+    numArgs = 0;
+    int beginIdx = 0;
+    int idx = command.indexOf(" ");
+
+    String arg;
+    char charBuffer[16];
+
+    while (idx != -1)
+    {
+        arg = command.substring(beginIdx, idx);
+
+        // add error handling for atoi:
+        args[numArgs++] = arg;
+        beginIdx = idx + 1;
+        idx = command.indexOf(" ", beginIdx);
+    }
+
+    arg = command.substring(beginIdx);
+    args[numArgs++] = arg;
+}
+
+/* This routine checks for any input waiting on the serial line. If any is
+ * available it is read in and added to a 128 character buffer. It sends back
+ * an error should the buffer overflow, and starts overwriting the buffer
+ * at that point. It only reads one character per call. If it receives a
+ * newline character is then runs the parseAndExecuteCommand() routine.
+ */
+void checkInput() {
+    int inbyte;
+    static char incomingBuffer[128];
+    static char bufPosition=0;
+
+    if(Serial.available()>0) {
+        // Read only one character per call
+        inbyte = Serial.read();
+        if(inbyte==10||inbyte==13) {
+            // Newline detected
+            incomingBuffer[bufPosition]='\0'; // NULL terminate the string
+            bufPosition=0; // Prepare for next command
+
+            // Supply a separate routine for parsing the command. This will
+            // vary depending on the task.
+            parseAndExecuteCommand(String(incomingBuffer));
+        }
+        else {
+            incomingBuffer[bufPosition]=(char)inbyte;
+            bufPosition++;
+            if(bufPosition==128) {
+                Serial.println("error: command overflow");
+                bufPosition=0;
+            }
+        }
+    }
+}
+
+/* This routine parses and executes any command received. It will have to be
+ * rewritten for any sketch to use the appropriate commands and arguments for
+ * the program you design. I find it easier to separate the input assembly
+ * from parsing so that I only have to modify this function and can keep the
+ * checkInput() function the same in each sketch.
+ */
+void parseAndExecuteCommand(String command) {
+    Serial.println("> " + command);
+    parse_args(command);
+    if(args[0].equals(String("ping"))) {
+        if(numArgs == 1) {
+            Serial.println("ok");
+        } else {
+            Serial.println("error: usage - 'ping'");
+        }
+    }
+    else if(args[0].equals(String("le"))) { // led set
+        if(numArgs == 2) {
+            if(args[1].equals(String("on"))) {
+                ledState = HIGH;
+                digitalWrite(LED,HIGH);
+                Serial.println("ok");
+            } else if(args[1].equals(String("off"))) {
+                ledState = LOW;
+                digitalWrite(LED,LOW);
+                Serial.println("ok");
+            } else {
+                Serial.println("error: usage - 'le [on/off]'");
+            }
+        } else {
+            Serial.println("error: usage - 'le [on/off]'");
+        }
+    }
+    else if(args[0].equals(String("rl"))) { // read led
+        if(numArgs == 1) {
+            Serial.println(ledState);
+        } else {
+            Serial.println("error: usage - 'rl'");
+        }
+    }
+    else if(args[0].equals(String("rs"))){ // read switches
+        if(numArgs == 2){
+            int indexNum = args[1].toInt();
+            if(indexNum > -1 && indexNum < 1){
+                Serial.println(digitalRead(switches[indexNum]));
+            } else {
+                Serial.println("Error: usage - rs [id]");
+            }
+        } else {
+            Serial.println("Error: usage - rs [id]");
+        }
+    }
+    else if(args[0].equals(String("ver"))) { // version information
+        if(numArgs == 1) {
+            String out = "Source last modified: ";
+            out += __TIMESTAMP__;
+            out += "\r\nPreprocessor timestamp: " __DATE__ " " __TIME__;
+            out += "\r\nSource code line number: ";
+            out += __LINE__;
+            out += "\r\nUsername: " STR(__USER__);
+            out += "\r\nDirectory: " STR(__DIR__);
+            out += "\r\nGit hash: " STR(__GIT_HASH__);
+            Serial.println(out);
+        } else {
+            Serial.println("error: usage - 'ver'");
+        }
+    }
+    else {
+        // Unrecognized command
+        Serial.println("error: unrecognized command");
+    }
+}
+
+double toDouble(String s)
+{
+  char buf[s.length() + 1];
+  s.toCharArray(buf, s.length() + 1);
+  return atof(buf);
+}
