+++ uno/uno.json
@@ -1,0 +1,1 @@
+[{"label":"front_right_bumper","type":"button","pin":13,"pullup":false,"invert":false}]
